Añadido validaciones

+++ src/main/java/org/repaso/dao/ComercialDAOImpl.java
@@ -38,7 +38,7 @@ public class ComercialDAOImpl implements ComercialDAO {
             ps.setString(idx++, comercial.getNombre());
             ps.setString(idx++, comercial.getApellido1());
             ps.setString(idx++, comercial.getApellido2());
-            ps.setFloat(idx, comercial.getComision());
+            ps.setBigDecimal(idx, comercial.getComision());
             return ps;
         },keyHolder);
 
@@ -55,7 +55,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                         rs.getString("nombre"),
                         rs.getString("apellido1"),
                         rs.getString("apellido2"),
-                        rs.getFloat("comisión"))
+                        rs.getBigDecimal("comisión"))
         );
 
         log.info("Devueltos {} registros.", listComercial.size());
@@ -71,7 +71,7 @@ public class ComercialDAOImpl implements ComercialDAO {
                                 rs.getString("nombre"),
                                 rs.getString("apellido1"),
                                 rs.getString("apellido2"),
-                                rs.getFloat("comisión"))
+                                rs.getBigDecimal("comisión"))
                         , id
                 );
